Fully functional* version of SW API completed
The internet today specifically is being glorious, so the API calls are
taking seconds EACH, if they work at all. As a result, while the JS
which was working fine as of yesterday, performs..... poorly. And errors
do pop in some of the calls. This could be improved by redesigning the
app to load everything into state, so it need only be called once. I
considered this architecture for some time, loading it all into an
array, however ultimately the point of this project was to play with
API calls so went with this architecture instead.