Additional updates to Star Wars module
SWapi.dev is still running slowly today, but at least it is functioning.
(I cross checked with other people's use of the API, and they too are
running poorly. Generally I assume if something isn't working well it is
my fault, but in this case it at least appears it might not be.) This
does however make my choice of calling the API more times than
absolutely necessary look like a bad decision. With the benefit of
hindsight, I should have loaded everything into state up front, so the
user experience could be crisper. It was so little data. At this point
though, I'd have to significantly rework the design to do so. If this were
for a client I would insist on doing so, but as this was just a learning
experience for playing with APIs, I think the requirement has been met.